Add smex-ido-M-x and smex-simple-M-x instead of old smex command
Initially smex was designed as an alternative to the excellent
IDO-based M-x(execute-extended-command) command. However, there is
actually a demand for another complementary interface like helm-smex.
These days, vertico and selectrum are examples, but they replace the
completing-read command, so you don't need to create a separate
package like helm-smex.

Add smex-M-x, smex-ido-M-x and smex-simple-M-x in this commit.
Traditional smex users believe that smex depends on IDO, so it's an
alias for smex-ido-M-x. By binding the key to smex-M-x, the new user
will automatically select either IDO or completing-read according to
the state of ido-mode.